add reserved_clients to listen section (#1553)
If the global client_max is reached, at most
reserved_clients connections can be opened against
the port specified in listen section

+++ test/functional/tests/client_max/runner.sh
@@ -13,9 +13,25 @@ if psql 'host=localhost port=6432 user=postgres dbname=postgres sslmode=disable'
 	exit 1
 fi
 
+# reserve_clients for listen allows to add some connections
+psql 'host=localhost port=9999 user=postgres dbname=postgres sslmode=disable' -c 'select pg_sleep(10000)' &
+p1=$!
+
+psql 'host=localhost port=9999 user=postgres dbname=postgres sslmode=disable' -c 'select pg_sleep(10000)' &
+p2=$!
+
+if psql 'host=localhost port=9999 user=postgres dbname=postgres sslmode=disable' -c 'select 42'; then
+	echo "the connect must fail"
+	exit 1
+fi
+
+kill -9 $p2
+kill -9 $p1
+
 kill -9 $p
 sleep 0.1
 
 psql 'host=localhost port=6432 user=postgres dbname=postgres sslmode=disable' -c 'select 42'
+psql 'host=localhost port=9999 user=postgres dbname=postgres sslmode=disable' -c 'select 42'
 
 ody-stop
